buildCodeJson done
Doesn't look great, but hey, it works!
Used delegates & some other tricks to avoid code repetition. 
The biggest problem with building json low-memory style is that you have to do it twice, first time to get the content length, second time to send it to the server. I don't have the memory to store the whole thing in-memory.