Menu skeleton
Screen has to be implemented -> it's gonna hold information about what is currently on-screen & manipulate 
data that has to be on it.

Menus contain items, setting & submenus.
Currently, Item is the backbone of the the whole system. It is a menu item that can't open anything. It serves a limited purpose as a title, separator, additional info, etc.
Menu is an Item that holds other items inside of itself. It differs from an Item because it can be pressed and will open a submenu.
I achieved (through Variadic functions) addition of undefined number of items at once. It adds objects recursively and tries to cast it to an item. All data is always saved as an Item, but you can cast it and expose additional functions by checking it's type (ITEM, SETTING, MENU).

Setting is an Item that performs a certain Action. Action will be a delegate function (or a set of delegate functions) that does something in the system on command. There will be multiple flavors of settings, ranging from simple boolean setting, to value setting etc.