Binding complete
TODO:
1) add more promises (a lot of them are missing currently)
2) update tempValues when automatically switching bits off. (this WILL
be hard, it'll require some sort of a callback system that registers
settings connected to that bit.
3) Promises have to return boolean, although this is offers much greater
flexibility (templates are an option)
4) Bind Remote to Actions (through registering Actions directly in
Remote class (we already have a good system for that, or reusing actions
ActionBinaryTree from Control. This would mean we'd have to include int
remoteCode in Action and then add some sort of retrieve for remoteCode.
I don't like that option since its' more of a hack than a real solution.
I'd like a real 1:1 binding, possibly RemoteAction class that contains
an action and a code, and then extending it with a BinaryTree template.)